Fix a truckload of clippy annotations, warnings and errors.

+++ src/log.rs
@@ -13,17 +13,19 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-use std::collections::{VecDeque, HashMap};
-use std::io::{Read, Write, Seek};
-use std::convert::TryInto;
-use std::sync::atomic::{AtomicBool, AtomicU64, AtomicU32, Ordering};
-use parking_lot::{Condvar, Mutex, RwLock, RwLockWriteGuard, MappedRwLockWriteGuard};
 use crate::{
-	error::{Error, Result},
-	table::TableId as ValueTableId,
-	index::{TableId as IndexTableId, Chunk as IndexChunk, ENTRY_BYTES},
-	options::Options,
 	column::ColId,
+	error::{Error, Result},
+	index::{Chunk as IndexChunk, TableId as IndexTableId, ENTRY_BYTES},
+	options::Options,
+	table::TableId as ValueTableId,
+};
+use parking_lot::{Condvar, MappedRwLockWriteGuard, Mutex, RwLock, RwLockWriteGuard};
+use std::{
+	collections::{HashMap, VecDeque},
+	convert::TryInto,
+	io::{Read, Seek, Write},
+	sync::atomic::{AtomicBool, AtomicU32, AtomicU64, Ordering},
 };
 
 const MAX_LOG_POOL_SIZE: usize = 16;
@@ -52,8 +54,13 @@ pub enum LogAction {
 }
 
 pub trait LogQuery {
-	fn with_index<R, F: FnOnce(&IndexChunk) -> R> (&self, table: IndexTableId, index: u64, f: F) -> Option<R>;
-	fn value(&self, table: ValueTableId, index: u64, dest: &mut[u8]) -> bool;
+	fn with_index<R, F: FnOnce(&IndexChunk) -> R>(
+		&self,
+		table: IndexTableId,
+		index: u64,
+		f: F,
+	) -> Option<R>;
+	fn value(&self, table: ValueTableId, index: u64, dest: &mut [u8]) -> bool;
 }
 
 #[derive(Default)]
@@ -70,23 +77,36 @@ impl LogOverlays {
 }
 
 impl LogQuery for RwLock<LogOverlays> {
-	fn with_index<R, F: FnOnce(&IndexChunk) -> R> (&self, table: IndexTableId, index: u64, f: F) -> Option<R> {
+	fn with_index<R, F: FnOnce(&IndexChunk) -> R>(
+		&self,
+		table: IndexTableId,
+		index: u64,
+		f: F,
+	) -> Option<R> {
 		self.read().with_index(table, index, f)
 	}
 
-	fn value(&self, table: ValueTableId, index: u64, dest: &mut[u8]) -> bool {
+	fn value(&self, table: ValueTableId, index: u64, dest: &mut [u8]) -> bool {
 		self.read().value(table, index, dest)
 	}
 }
 
 impl LogQuery for LogOverlays {
-	fn with_index<R, F: FnOnce(&IndexChunk) -> R> (&self, table: IndexTableId, index: u64, f: F) -> Option<R> {
-		self.index.get(&table).and_then(|o| o.map.get(&index).map(|(_id, _mask, data)| f(data)))
+	fn with_index<R, F: FnOnce(&IndexChunk) -> R>(
+		&self,
+		table: IndexTableId,
+		index: u64,
+		f: F,
+	) -> Option<R> {
+		self.index
+			.get(&table)
+			.and_then(|o| o.map.get(&index).map(|(_id, _mask, data)| f(data)))
 	}
 
-	fn value(&self, table: ValueTableId, index: u64, dest: &mut[u8]) -> bool {
+	fn value(&self, table: ValueTableId, index: u64, dest: &mut [u8]) -> bool {
 		let s = self;
-		if let Some(d) = s.value.get(&table).and_then(|o| o.map.get(&index).map(|(_id, data)| data)) {
+		if let Some(d) = s.value.get(&table).and_then(|o| o.map.get(&index).map(|(_id, data)| data))
+		{
 			let len = dest.len().min(d.len());
 			dest[0..len].copy_from_slice(&d[0..len]);
 			true
@@ -152,7 +172,7 @@ impl<'a> LogReader<'a> {
 		let mut buf = [0u8; 8];
 		read_buf(1, &mut buf)?;
 		match buf[0] {
-			BEGIN_RECORD =>  {
+			BEGIN_RECORD => {
 				read_buf(8, &mut buf)?;
 				let record_id = u64::from_le_bytes(buf);
 				self.record_id = record_id;
@@ -160,7 +180,8 @@ impl<'a> LogReader<'a> {
 			},
 			INSERT_INDEX => {
 				read_buf(2, &mut buf)?;
-				let table = IndexTableId::from_u16(u16::from_le_bytes(buf[0..2].try_into().unwrap()));
+				let table =
+					IndexTableId::from_u16(u16::from_le_bytes(buf[0..2].try_into().unwrap()));
 				read_buf(8, &mut buf)?;
 				let index = u64::from_le_bytes(buf);
 				self.cleared.index.push((table, index));
@@ -168,7 +189,8 @@ impl<'a> LogReader<'a> {
 			},
 			INSERT_VALUE => {
 				read_buf(2, &mut buf)?;
-				let table = ValueTableId::from_u16(u16::from_le_bytes(buf[0..2].try_into().unwrap()));
+				let table =
+					ValueTableId::from_u16(u16::from_le_bytes(buf[0..2].try_into().unwrap()));
 				read_buf(8, &mut buf)?;
 				let index = u64::from_le_bytes(buf);
 				self.cleared.values.push((table, index));
@@ -195,12 +217,11 @@ impl<'a> LogReader<'a> {
 			},
 			DROP_TABLE => {
 				read_buf(2, &mut buf)?;
-				let table = IndexTableId::from_u16(u16::from_le_bytes(buf[0..2].try_into().unwrap()));
+				let table =
+					IndexTableId::from_u16(u16::from_le_bytes(buf[0..2].try_into().unwrap()));
 				Ok(LogAction::DropTable(table))
 			},
-			_ => {
-				Err(Error::Corruption("Bad log entry type".into()))
-			},
+			_ => Err(Error::Corruption("Bad log entry type".into())),
 		}
 	}
 
@@ -230,9 +251,7 @@ pub struct LogChange {
 }
 
 impl LogChange {
-	fn new(
-		record_id: u64,
-	) -> LogChange {
+	fn new(record_id: u64) -> LogChange {
 		LogChange {
 			local_index: Default::default(),
 			local_values: Default::default(),
@@ -245,14 +264,16 @@ impl LogChange {
 		self.local_values.get(&id)
 	}
 
-	pub fn to_file(self, file: &mut std::io::BufWriter<std::fs::File>)
-		-> Result<(HashMap<IndexTableId, IndexLogOverlay>, HashMap<ValueTableId, ValueLogOverlay>, u64)>
+	pub fn to_file(
+		self,
+		file: &mut std::io::BufWriter<std::fs::File>,
+	) -> Result<(HashMap<IndexTableId, IndexLogOverlay>, HashMap<ValueTableId, ValueLogOverlay>, u64)>
 	{
 		let mut crc32 = crc32fast::Hasher::new();
 		let mut bytes: u64 = 0;
 
 		let mut write = |buf: &[u8]| -> Result<()> {
-			file.write(buf)?;
+			file.write_all(buf)?;
 			crc32.update(buf);
 			bytes += buf.len() as u64;
 			Ok(())
@@ -263,21 +284,21 @@ impl LogChange {
 
 		for (id, overlay) in self.local_index.iter() {
 			for (index, (_, modified_entries_mask, chunk)) in overlay.map.iter() {
-				write(&INSERT_INDEX.to_le_bytes().as_ref())?;
+				write(INSERT_INDEX.to_le_bytes().as_ref())?;
 				write(&id.as_u16().to_le_bytes())?;
 				write(&index.to_le_bytes())?;
 				write(&modified_entries_mask.to_le_bytes())?;
 				let mut mask = *modified_entries_mask;
 				while mask != 0 {
 					let i = mask.trailing_zeros();
-					mask = mask & !(1 << i);
-					write(&chunk[i as usize *ENTRY_BYTES .. (i as usize + 1)*ENTRY_BYTES])?;
+					mask &= !(1 << i);
+					write(&chunk[i as usize * ENTRY_BYTES..(i as usize + 1) * ENTRY_BYTES])?;
 				}
 			}
 		}
 		for (id, overlay) in self.local_values.iter() {
 			for (index, (_, value)) in overlay.map.iter() {
-				write(&INSERT_VALUE.to_le_bytes().as_ref())?;
+				write(INSERT_VALUE.to_le_bytes().as_ref())?;
 				write(&id.as_u16().to_le_bytes())?;
 				write(&index.to_le_bytes())?;
 				write(value)?;
@@ -285,12 +306,12 @@ impl LogChange {
 		}
 		for id in self.dropped_tables.iter() {
 			log::debug!(target: "parity-db", "Finalizing drop {}", id);
-			write(&DROP_TABLE.to_le_bytes().as_ref())?;
+			write(DROP_TABLE.to_le_bytes().as_ref())?;
 			write(&id.as_u16().to_le_bytes())?;
 		}
 		write(&END_RECORD.to_le_bytes())?;
 		let checksum: u32 = crc32.finalize();
-		file.write(&checksum.to_le_bytes())?;
+		file.write_all(&checksum.to_le_bytes())?;
 		bytes += 4;
 		file.flush()?;
 		Ok((self.local_index, self.local_values, bytes))
@@ -303,14 +324,8 @@ pub struct LogWriter<'a> {
 }
 
 impl<'a> LogWriter<'a> {
-	pub fn new(
-		overlays: &'a RwLock<LogOverlays>,
-		record_id: u64,
-	) -> LogWriter<'a> {
-		LogWriter {
-			overlays,
-			log: LogChange::new(record_id),
-		}
+	pub fn new(overlays: &'a RwLock<LogOverlays>, record_id: u64) -> LogWriter<'a> {
+		LogWriter { overlays, log: LogChange::new(record_id) }
 	}
 
 	pub fn record_id(&self) -> u64 {
@@ -320,16 +335,21 @@ impl<'a> LogWriter<'a> {
 	pub fn insert_index(&mut self, table: IndexTableId, index: u64, sub: u8, data: &IndexChunk) {
 		match self.log.local_index.entry(table).or_default().map.entry(index) {
 			std::collections::hash_map::Entry::Occupied(mut entry) => {
-				*entry.get_mut() = (self.log.record_id, entry.get().1 | (1 << sub), data.clone());
-			}
+				*entry.get_mut() = (self.log.record_id, entry.get().1 | (1 << sub), *data);
+			},
 			std::collections::hash_map::Entry::Vacant(entry) => {
-				entry.insert((self.log.record_id, 1 << sub, data.clone()));
-			}
+				entry.insert((self.log.record_id, 1 << sub, *data));
+			},
 		}
 	}
 
 	pub fn insert_value(&mut self, table: ValueTableId, index: u64, data: Vec<u8>) {
-		self.log.local_values.entry(table).or_default().map.insert(index, (self.log.record_id, data.clone()));
+		self.log
+			.local_values
+			.entry(table)
+			.or_default()
+			.map
+			.insert(index, (self.log.record_id, data));
 	}
 
 	pub fn drop_table(&mut self, id: IndexTableId) {
@@ -342,15 +362,30 @@ impl<'a> LogWriter<'a> {
 }
 
 impl<'a> LogQuery for LogWriter<'a> {
-	fn with_index<R, F: FnOnce(&IndexChunk) -> R> (&self, table: IndexTableId, index: u64, f: F) -> Option<R> {
-		match self.log.local_index.get(&table).and_then(|o| o.map.get(&index).map(|(_id, _mask, data)| data)) {
+	fn with_index<R, F: FnOnce(&IndexChunk) -> R>(
+		&self,
+		table: IndexTableId,
+		index: u64,
+		f: F,
+	) -> Option<R> {
+		match self
+			.log
+			.local_index
+			.get(&table)
+			.and_then(|o| o.map.get(&index).map(|(_id, _mask, data)| data))
+		{
 			Some(data) => Some(f(data)),
 			None => self.overlays.with_index(table, index, f),
 		}
 	}
 
-	fn value(&self, table: ValueTableId, index: u64, dest: &mut[u8]) -> bool {
-		if let Some(d) = self.log.local_values.get(&table).and_then(|o| o.map.get(&index).map(|(_id, data)| data)) {
+	fn value(&self, table: ValueTableId, index: u64, dest: &mut [u8]) -> bool {
+		if let Some(d) = self
+			.log
+			.local_values
+			.get(&table)
+			.and_then(|o| o.map.get(&index).map(|(_id, data)| data))
+		{
 			let len = dest.len().min(d.len());
 			dest[0..len].copy_from_slice(&d[0..len]);
 			true
@@ -366,18 +401,42 @@ pub struct IdentityHash(u64);
 pub type BuildIdHash = std::hash::BuildHasherDefault<IdentityHash>;
 
 impl std::hash::Hasher for IdentityHash {
-    fn write(&mut self, _: &[u8]) { unreachable!() }
-    fn write_u8(&mut self, _: u8)       { unreachable!() }
-    fn write_u16(&mut self, _: u16)     { unreachable!() }
-    fn write_u32(&mut self, _: u32)     { unreachable!() }
-    fn write_u64(&mut self, n: u64)     { self.0 = n }
-    fn write_usize(&mut self, _: usize) { unreachable!() }
-    fn write_i8(&mut self, _: i8)       { unreachable!() }
-    fn write_i16(&mut self, _: i16)     { unreachable!() }
-    fn write_i32(&mut self, _: i32)     { unreachable!() }
-    fn write_i64(&mut self, _: i64)     { unreachable!() }
-    fn write_isize(&mut self, _: isize) { unreachable!() }
-    fn finish(&self) -> u64 { self.0 }
+	fn write(&mut self, _: &[u8]) {
+		unreachable!()
+	}
+	fn write_u8(&mut self, _: u8) {
+		unreachable!()
+	}
+	fn write_u16(&mut self, _: u16) {
+		unreachable!()
+	}
+	fn write_u32(&mut self, _: u32) {
+		unreachable!()
+	}
+	fn write_u64(&mut self, n: u64) {
+		self.0 = n
+	}
+	fn write_usize(&mut self, _: usize) {
+		unreachable!()
+	}
+	fn write_i8(&mut self, _: i8) {
+		unreachable!()
+	}
+	fn write_i16(&mut self, _: i16) {
+		unreachable!()
+	}
+	fn write_i32(&mut self, _: i32) {
+		unreachable!()
+	}
+	fn write_i64(&mut self, _: i64) {
+		unreachable!()
+	}
+	fn write_isize(&mut self, _: isize) {
+		unreachable!()
+	}
+	fn finish(&self) -> u64 {
+		self.0
+	}
 }
 
 #[derive(Default)]
@@ -457,7 +516,7 @@ impl Log {
 				}
 			}
 		}
-		logs.make_contiguous().sort_by_key(|(_id, record_id,  _)| *record_id);
+		logs.make_contiguous().sort_by_key(|(_id, record_id, _)| *record_id);
 		let next_log_id = if logs.is_empty() { 0 } else { max_log_id + 1 };
 
 		Ok(Log {
@@ -491,7 +550,7 @@ impl Log {
 	pub fn open_log_file(path: &std::path::Path) -> Result<(std::fs::File, Option<u64>)> {
 		let mut file = std::fs::OpenOptions::new().read(true).write(true).open(path)?;
 		if file.metadata()?.len() == 0 {
-			return Ok((file, None));
+			return Ok((file, None))
 		}
 		// read first record id
 		let mut buf = [0; 9];
@@ -534,13 +593,9 @@ impl Log {
 		Ok(())
 	}
 
-	pub fn begin_record<'a>(&'a self) -> LogWriter<'a> {
+	pub fn begin_record(&self) -> LogWriter<'_> {
 		let id = self.next_record_id.fetch_add(1, Ordering::Relaxed);
-		let writer = LogWriter::new(
-			&self.overlays,
-			id
-		);
-		writer
+		LogWriter::new(&self.overlays, id)
 	}
 
 	pub fn end_record(&self, log: LogChange) -> Result<u64> {
@@ -556,15 +611,12 @@ impl Log {
 				// find a free id
 				let id = self.next_log_id.fetch_add(1, Ordering::SeqCst);
 				let path = Self::log_path(&self.path, id);
-				let file = std::fs::OpenOptions::new().create(true).read(true).write(true).open(path)?;
+				let file =
+					std::fs::OpenOptions::new().create(true).read(true).write(true).open(path)?;
 				log::debug!(target: "parity-db", "Flush: Activated new writer {}", id);
 				(id, file)
 			};
-			*appending = Some(Appending {
-				size: 0,
-				file: std::io::BufWriter::new(file),
-				id,
-			});
+			*appending = Some(Appending { size: 0, file: std::io::BufWriter::new(file), id });
 		}
 		let appending = appending.as_mut().unwrap();
 		let (index, values, bytes) = log.to_file(&mut appending.file)?;
@@ -600,25 +652,19 @@ impl Log {
 		let mut overlays = self.overlays.write();
 		for (table, index) in cleared.index.into_iter() {
 			if let Some(ref mut overlay) = overlays.index.get_mut(&table) {
-				match overlay.map.entry(index) {
-					std::collections::hash_map::Entry::Occupied(e) => {
-						if e.get().0 == record_id {
-							e.remove_entry();
-						}
+				if let std::collections::hash_map::Entry::Occupied(e) = overlay.map.entry(index) {
+					if e.get().0 == record_id {
+						e.remove_entry();
 					}
-					_ => {},
 				}
 			}
 		}
 		for (table, index) in cleared.values.into_iter() {
 			if let Some(ref mut overlay) = overlays.value.get_mut(&table) {
-				match overlay.map.entry(index) {
-					std::collections::hash_map::Entry::Occupied(e) => {
-						if e.get().0 == record_id {
-							e.remove_entry();
-						}
+				if let std::collections::hash_map::Entry::Occupied(e) = overlay.map.entry(index) {
+					if e.get().0 == record_id {
+						e.remove_entry();
 					}
-					_ => {},
 				}
 			}
 		}
@@ -634,7 +680,7 @@ impl Log {
 		if flushing.is_some() {
 			let mut reading_state = self.reading_state.lock();
 
-			while *reading_state == ReadingState::Reading  {
+			while *reading_state == ReadingState::Reading {
 				log::debug!(target: "parity-db", "Flush: Awaiting log reader");
 				self.done_reading_cv.wait(&mut reading_state)
 			}
@@ -698,10 +744,7 @@ impl Log {
 		}
 		if let Some((id, _record_id, file)) = self.replay_queue.write().pop_front() {
 			log::debug!(target: "parity-db", "Replay: Activated log reader {}", id);
-			*reading = Some(Reading {
-				id,
-				file: std::io::BufReader::new(file),
-			});
+			*reading = Some(Reading { id, file: std::io::BufReader::new(file) });
 			*self.reading_state.lock() = ReadingState::Reading;
 			Ok(Some(id))
 		} else {
@@ -711,9 +754,7 @@ impl Log {
 	}
 
 	pub fn clean_logs(&self, count: usize) -> Result<bool> {
-		let mut cleaned: Vec<_> = {
-			self.cleanup_queue.write().drain(0..count).collect()
-		};
+		let mut cleaned: Vec<_> = { self.cleanup_queue.write().drain(0..count).collect() };
 		for (id, ref mut file) in cleaned.iter_mut() {
 			log::debug!(target: "parity-db", "Cleaned: {}", id);
 			file.seek(std::io::SeekFrom::Start(0))?;
@@ -738,33 +779,31 @@ impl Log {
 		self.cleanup_queue.read().len()
 	}
 
-	pub fn read_next<'a>(&'a self, validate: bool) -> Result<Option<LogReader<'a>>> {
+	pub fn read_next(&self, validate: bool) -> Result<Option<LogReader<'_>>> {
 		let mut reading_state = self.reading_state.lock();
 		if *reading_state != ReadingState::Reading {
 			log::trace!(target: "parity-db", "No logs to enact");
-			return Ok(None);
+			return Ok(None)
 		}
 
 		let reading = self.reading.write();
 		if reading.is_none() {
 			log::trace!(target: "parity-db", "No active reader");
-			return Ok(None);
+			return Ok(None)
 		}
 		let reading = RwLockWriteGuard::map(reading, |r| &mut r.as_mut().unwrap().file);
 		let mut reader = LogReader::new(reading, validate);
 		match reader.next() {
-			Ok(LogAction::BeginRecord) => {
-				return Ok(Some(reader));
-			}
-			Ok(_) => return Err(Error::Corruption("Bad log record structure".into())),
+			Ok(LogAction::BeginRecord) => Ok(Some(reader)),
+			Ok(_) => Err(Error::Corruption("Bad log record structure".into())),
 			Err(Error::Io(e)) if e.kind() == std::io::ErrorKind::UnexpectedEof => {
 				*reading_state = ReadingState::Idle;
 				self.done_reading_cv.notify_one();
 				log::debug!(target: "parity-db", "Read: End of log");
-				return Ok(None);
-			}
-			Err(e) => return Err(e),
-		};
+				Ok(None)
+			},
+			Err(e) => Err(e),
+		}
 	}
 
 	pub fn overlays(&self) -> &RwLock<LogOverlays> {
